ut: add support for fs_type = "none"

+++ src/test/unittest/unittest.sh
@@ -120,6 +120,9 @@ else
 	non-pmem)
 		DIR=$NON_PMEM_FS_DIR/$curtestdir$UNITTEST_NUM
 		;;
+	none)
+		DIR=/dev/null/not_existing_dir/$curtestdir$UNITTEST_NUM
+		;;
 	esac
 	[ "$DIR" ] || {
 		[ "$UNITTEST_QUIET" ] || echo "$UNITTEST_NAME: SKIP fs-type $FS (not configured)"
@@ -770,6 +773,11 @@ function setup() {
 		exit 0
 	fi
 
+	# fs type "none" must be explicitly enabled
+	if [ "$FS" = "none" -a "$req_fs_type" != "1" ]; then
+		exit 0
+	fi
+
 	if [ "$MEMCHECK" = "force-enable" ]; then
 		export RUN_MEMCHECK=1
 	fi
@@ -784,10 +792,13 @@ function setup() {
 
 	rm -f check_pool_${BUILD}_${UNITTEST_NUM}.log
 
-	if [ -d "$DIR" ]; then
-		rm --one-file-system -rf -- $DIR
+	if [ "$FS" != "none" ]; then
+		if [ -d "$DIR" ]; then
+			rm --one-file-system -rf -- $DIR
+		fi
+
+		mkdir $DIR
 	fi
-	mkdir $DIR
 }
 
 #
@@ -804,7 +815,9 @@ function pass() {
 	msg="PASS"
 	[ -t 1 ] && command -v tput >/dev/null && msg="$(tput setaf 2)$msg$(tput sgr0)"
 	echo -e "$UNITTEST_NAME: $msg"
-	rm --one-file-system -rf -- $DIR
+	if [ "$FS" != "none" ]; then
+		rm --one-file-system -rf -- $DIR
+	fi
 }
 
 # Length of pool file's signature
